refactor: move renameColumnQuery to a separate file
